requirements.txt
python-dotenv
python-telegram-bot
redis
requests